tests/submission/Infraware/Forms/contents/Forms/output_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_form</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <output name=o id='output_id'></output>
	</form>

	<script>
		
		var output = document.getElementById("output_id");
		form = output.form;

		if (typeof(form) == "object") {
			test(function() {
				assert_equals(form.name, "input_form", "form attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("form attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
